refactor: - BigDecimal for price - PayPal payment method added - public mapper method

+++ src/main/java/org/example/petstore/dto/ReceiptDto.java
@@ -15,7 +15,7 @@ public class ReceiptDto {
     private String paymentMethod;
     private String status;
     private List<OrderLineDto> orderLines;
-    private double totalAmount;
+    private BigDecimal totalAmount;
 
     @Data
     public static class OrderLineDto {
